Wrap all React tests with `describeEachReactMode`
